SEM15LUISCALLE: Restructuracion de ConductorEntity
