+++ src/main/java/com/transport/management/modules/conductor/ConductorController.java
@@ -12,7 +12,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.transport.management.entities.ConductorEntity;
 import com.transport.management.services.ConductorService;
 
 @RestController
