+++ src/main/java/com/transport/management/repositories/ConductorRepository.java
@@ -2,7 +2,7 @@ package com.transport.management.repositories;
 
 import java.util.List;
 
-import com.transport.management.entities.ConductorEntity;
+import com.transport.management.modules.conductor.ConductorEntity;
 import com.transport.management.utils.abtractBase.BaseRepository;
 
 public interface ConductorRepository extends BaseRepository<ConductorEntity> {
